# Optimized vec::IntoIter::next_chunk impl