@Mark-Simulacrum 2022-07-26
r=me

@bors rollup=never in case has unexpected perf effects